Add files via upload
This is a 2024 update to the original LA 83 majors escape room game created in 2023. This update includes accessibility considerations (fonts and colors, size), updated descriptions for clues, some clues were revised entirely, and the main branches (intro, ending) were updated to be more general across any sections of the course, for any instructor who desires to use it.